Move JwtUtils tests to separate file.
